Move UserInputError to its own file
