+++ packages/server/src/schema/connection/types.ts
@@ -1,5 +1,3 @@
-import { GraphQLError } from 'graphql'
-
 import { PageInfo } from '../../types'
 
 export interface Connection<T> {
@@ -14,14 +12,4 @@ export interface ConnectionPayload {
   before?: string
   first?: number
   last?: number
-}
-
-export class UserInputError extends GraphQLError {
-  constructor(message: string, code: string) {
-    super(message, {
-      extensions: {
-        code: code,
-      },
-    })
-  }
 }
